requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
# EDA Diário (rota /eda/): ajuste versões se necessário
mysql-connector-python>=8.0.0
pandas>=2.0.0
openpyxl>=3.1.0
